#50 fix warnings

+++ test/ChaiVM/interpreter/code-manager-test.cpp
@@ -37,7 +37,7 @@ TEST(CodeManager, returnsBytecodes) {
     }
 
     const chai::chsize_t initial_pc = manager.startPC();
-    for (int i = 0; i < seq.size(); ++i) {
+    for (size_t i = 0; i < seq.size(); ++i) {
         chai::chsize_t pc = initial_pc + i * sizeof(chai::bytecode_t);
         EXPECT_EQ(manager.getBytecode(pc), seq[i]);
     }
